Fix int32 overflow in selected_rows directory (3 files)

+++ paddle/phi/kernels/selected_rows/gpu/adam_kernel.cu
@@ -236,8 +236,8 @@ void AdamDenseParamSparseGradKernel(
 
   if (beta1_pow.place() == CPUPlace() && beta2_pow.place() == CPUPlace()) {
     int threads = 512;
-    int ndim = param.numel();
-    int blocks = (ndim + threads - 1) / threads;
+    int64_t ndim = param.numel();
+    int64_t blocks = (ndim + threads - 1) / threads;
 
     SparseAdamCUDAKernelREG<T, MPDType>
         <<<blocks, threads, 0, dev_ctx.stream()>>>(
